Implement resource node management with locking mechanism
- Refactored the remote.py test to directly add names to the names node.
- Introduced a new node.py file to handle API endpoints for node information and lock status.
- Created a lock.py file to implement a read-write lock mechanism for resource nodes.
- Added an init method to set up the database for locks and ensure no stale locks exist.
- Developed a node management system in node.py to handle resource nodes, including mounting, unmounting, and proxying nodes.
- Implemented a template system for launching CRM servers with appropriate parameters.
- Enhanced the Treeger class to manage node relationships and database interactions.
- Added schemas for resource node information to facilitate data handling.
