+++ .env.example
@@ -1,7 +1,5 @@
-SERVER_PORT=8000_as_default
+SQLITE_PATH=/your/scene/path/here # default: ./noodle.db
 
-SQLITE_PATH=/your/scene/path/here
+MEMORY_TEMP_PATH=/your/temp/directory/here # default: ./temp
 
-MEMORY_TEMP_PATH=/your/temp/directory/here
-
-NOODLE_CONFIG_PATH=/your/noodle/config/path/here
+NOODLE_CONFIG_PATH=/your/noodle/config/path/here # default: ./noodle.config.yaml
