+++ src/pynoodle/__init__.py
@@ -1,4 +1,3 @@
-from .scene import ResourceNode
-from .noodle import Noodle, noodle
+from .noodle import noodle
+from .module_cache import ResourceNodeTemplate
 from .hooks import NOODLE_INIT, NOODLE_TERMINATE
-from .module_cache import ModuleCache, ResourceNodeTemplate
