+++ src/pynoodle/node/__init__.py
@@ -1,3 +1,3 @@
 from .lock import RWLock
 from .treeger import Treeger
-from .scene_node import ResourceNode
+from .node import ResourceNode
