+++ src/pynoodle/schemas/lock.py
@@ -2,3 +2,6 @@ from pydantic import BaseModel
 
 class LockInfo(BaseModel):
     lock_id: str
+
+class LockedInfo(BaseModel):
+    locked: bool
