+++ src/pynoodle/schemas/node.py
