--- tests/crms/hello/__init__.py
@@ -1,9 +1,0 @@
-from pynoodle import RawScenarioNode
-
-from tests.icrms.ihello import IHello
-from .hello import Hello
-
-RAW = RawScenarioNode(
-    CRM=Hello,
-    ICRM=IHello,
-)
