+++ tests/remote.py
@@ -19,11 +19,10 @@ async def lifespan(app: FastAPI):
 
     noodle.mount_node('nameSet', 'names')
     with noodle.connect_node(INames, 'nameSet', 'lw') as names:
-        crm = names.crm
-        crm.add_name('Alice')
-        crm.add_name('Bob')
-        crm.add_name('Charlie')
-    
+        names.add_name('Alice')
+        names.add_name('Bob')
+        names.add_name('Charlie')
+
     yield
     
     noodle.unmount_node('nameSet')
